clean up simulation's CMakeLists.txt and fix the build under Windows
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@5646 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ simulation/tools/sim_viewer.cpp
@@ -521,7 +521,7 @@ void wRo_to_euler(const Eigen::Matrix3f& wRo, double& yaw, double& pitch, double
   yaw = standardRad(atan2(wRo(1,0), wRo(0,0)));
   double c = cos(yaw);
   double s = sin(yaw);
-  pitch = standardRad(atan2(-wRo(2,0), wRo(0,0)*c + wRo(1,0)*s));
+  pitch = standardRad(atan2(static_cast<double> (-wRo(2,0)), wRo(0,0)*c + wRo(1,0)*s));
   roll  = standardRad(atan2(wRo(0,2)*s - wRo(1,2)*c, -wRo(0,1)*s + wRo(1,1)*c));
 }
